Add auth rules for the recipe show view

+++ app/helpers/recipes_helper.rb
@@ -27,10 +27,23 @@ module RecipesHelper
     content.html_safe
   end
 
+  def details_buttons(recipe)
+    content = ''
+    if can?(:create, recipe)
+      content << (link_to 'Generate shopping list', general_shopping_lists_url, class: 'buttons btn btn-primary auto-width').to_s
+      content << (link_to 'Add ingredient', new_recipe_food_path(recipe), class: 'buttons btn btn-primary auto-width').to_s
+    end
+    content.html_safe
+  end
+
   def recipe_actions(food, recipe_food)
     content = ''
-    content << (button_to 'Remove', recipe_food_path(food), method: :delete, class: 'btn btn-link link-warning').to_s
+    unless can?(:update, food) && can?(:destroy, food)
+      content << '<span>No actions permited</span>'
+      return content.html_safe
+    end
     content << (link_to 'Modify', edit_recipe_food_path(recipe_food), class: 'link-info p-2').to_s
+    content << (button_to 'Remove', recipe_food_path(food), method: :delete, class: 'btn btn-link link-warning').to_s
     content.html_safe
   end
 end
